Allow the normalization to be modified
In a conversation with @pmelchior it was suggested that normalizing
the morphology might not be necessary with the new ADAM algorithm in
scarlet, since it's better at breaking up degeneracies.
In order to test this on HSC data this commit added functionality
to rmove the normalization constraint. The result of the test was
that deblending takes longer with poorer results when not using
normalization, so this code will be stripped out in the next commit
but is stored here for historical purposes.