Fix inability to write unknown error type
Now that we trap unknown errors we still cannot write the new error
into the catalog because it would exceeed the character limit.
We could increase the number of characters in the field,
but since errors are rare this would result in a lot of wasted
disk space for very little gain, since unknown errors are so rare
that they are unlikely to happen once we begin running scarlet on
a regular basis and understand all of our errors
(even before they are all fixed).