Use mask plane for weights
At some point we stopped using the mask plane for weights and
started just using the inverse variance. There is some disagreement
over whether or not this is the right thing to do, and testing with
scarlet has shown including weights tends to produce worse results.
So this commit reintroduces the mask plane weights and makes that
the default configuration.