Turn off setSpectra initialization by default
For large skinny blends the `setSpectra` method of initialization
causes the initialization to run out of memory, so by default we
turn this option off. DM-28870 will investigate if there is a
criteria (such as peaks*bands*area) that we can use to determine
when it is safe to use setSpectra and when to skip that step.