Refactor ScarletDeblendTask
While implementing a fix to skip sky objects I noticed that
there are a number of places where columns were not being appropriately
updated. So this commit refactors ScarletDeblendTask to make it less
likely that similar errors can occur in the future, and also adds
tests to check that columns are being persisted properly.

The major change is that now all of the source updates are
occurring on the input (mergeDet) catalog, and only at the end
are the single band catalogs created and updated with band specific
information. An added benefit is that deblended sources have the
same source ID across catalogs, and are also now continuous catalogs.