Clip footprints to exposure
Deblended children near the edge of an image can have their
footprints dilated outside of the exposure by the PSF width.
This commit ensures that the dilated model is always contained
fully in the exposure. This is valid, as later in pipe_tasks
the science pipelines flag edge sources so it isn't expected
that all of their flux is contained in the image.