Revert to old modelToHeavy code with a slight change
The newer version of the code was creating the spans from the
scarlet model, dilating those spans, and using those spans
for the final footprints. The old version of the code simply
thresholded at zero to create the HeavyFootprints.

This commit reverts to the old code, with the change that it
performs the convolution and uses all pixels that are not zero
(including negative pixel values) in the footprint.

There is also a minor change to ensure that the MultibandFootprint
has all of the peaks, not just the single pand footprints.