Allow deblending to continue even when data is missing for some bands
Previously any footprints that could not construct a PSF in all
bands were skipped. As part of HSC PDR4, because there are patches
with only partial coverage of narrow bands (for example), we want
to allow the deblender to proceed.

This ticket implements deblending with only partial PSF coverage,
and also adds new columns to the catalog to keep track of the
coverage per band.