Properly check for max Footprint bounding Box area
The code currently has a threshold set for the maximum footprint
area, however in terms of our ability to process a field we
actually care about the bouding box that contains the footprint,
as that is directly proportional to the memory and runtime
required. This commit changes from the Footprint area to the
area of the Footprint's bounding Box.